feat(inputs.dpdk): move processCommand to dpdk_connector. Update identifySocket function. Remove prepareGlob func. Remove getArrayDiff func. Update tests
