+++ plugins/inputs/dpdk/dpdk_utils.go
@@ -11,7 +11,6 @@ import (
 	"strings"
 
 	"github.com/influxdata/telegraf/filter"
-	"github.com/influxdata/telegraf/internal/globpath"
 )
 
 func commandWithParams(command string, params string) string {
@@ -38,42 +37,6 @@ func getParams(command string) string {
 		return ""
 	}
 	return command[index+1:]
-}
-
-// getDiffArrays checks two arrays and return the diff between them
-// toAdd: return a list of the difference (set operation): newArray \ oldArray
-// toDelete: return a list of the difference (set operation): oldArray \ newArray
-func getDiffArrays(oldArray, newArray []string) (toAdd, toDelete []string) {
-	if len(oldArray) == 0 {
-		return newArray, nil
-	}
-
-	if len(newArray) == 0 {
-		return nil, oldArray
-	}
-
-	// Form maps
-	oldArrayMap := make(map[string]struct{})
-	for _, val := range oldArray {
-		oldArrayMap[val] = struct{}{}
-	}
-
-	// Form the list toAdd
-	newArrayMap := make(map[string]struct{})
-	for _, val := range newArray {
-		newArrayMap[val] = struct{}{}
-		if _, ok := oldArrayMap[val]; !ok {
-			toAdd = append(toAdd, val)
-		}
-	}
-
-	// Form the list toDelete
-	for key := range oldArrayMap {
-		if _, ok := newArrayMap[key]; !ok {
-			toDelete = append(toDelete, key)
-		}
-	}
-	return toAdd, toDelete
 }
 
 func (dpdk *dpdk) getDpdkInMemorySocketPaths() []string {
@@ -187,8 +150,4 @@ func uniqueValues(values []string) []string {
 
 func isEmpty(value interface{}) bool {
 	return value == nil || (reflect.ValueOf(value).Kind() == reflect.Ptr && reflect.ValueOf(value).IsNil())
-}
-
-func prepareGlob(path string) (*globpath.GlobPath, error) {
-	return globpath.Compile(path + "*")
 }
